# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "description": "Stable error shape per `docs/integration/02-gateway-rest-api.md`.\n\nAll non-2xx responses from `/api/v1/*` carry this body. The 11\n`error_code` values are the closed enum from\n`src/eveys_ocpp/api/_errors.py` (BAD_REQUEST, UNAUTHORIZED,\nFORBIDDEN, UNKNOWN_CP_ID, UNKNOWN_TRANSACTION_ID,\nUNKNOWN_RESERVATION_ID, CHARGER_OFFLINE, CHARGER_TIMEOUT,\nWINDOW_TOO_LARGE, RATE_LIMITED, INTERNAL_ERROR).",
        "example": {
          "error": "unknown cp_id: CP_DOES_NOT_EXIST",
          "error_code": "UNKNOWN_CP_ID",
          "request_id": "9b3c5d18-1f7c-4b6a-8e0e-5b9a3c4f2e10"
        },
        "properties": {
          "error": {
            "description": "Human-readable description for logs.",
            "title": "Error",
            "type": "string"
          },
          "error_code": {
            "description": "Stable machine code; one of the 11 closed-enum values.",
            "title": "Error Code",
            "type": "string"
          },
          "request_id": {
            "description": "Correlation id mirroring the X-Request-ID header.",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "error_code",
          "request_id"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Transaction": {
        "properties": {
          "connector_id": {
            "title": "Connector Id",
            "type": "integer"
          },
          "cp_id": {
            "title": "Cp Id",
            "type": "string"
          },
          "id_tag": {
            "description": "RFID / token used to authorize the session.",
            "title": "Id Tag",
            "type": "string"
          },
          "meter_start_wh": {
            "title": "Meter Start Wh",
            "type": "integer"
          },
          "meter_stop_wh": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meter Stop Wh"
          },
          "open": {
            "description": "True until StopTransaction lands.",
            "title": "Open",
            "type": "boolean"
          },
          "started_at": {
            "description": "ISO 8601 timestamp of the StartTransaction.",
            "title": "Started At",
            "type": "string"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Reason"
          },
          "stopped_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopped At"
          },
          "transaction_id": {
            "description": "OCPP 1.6 transaction id assigned by the gateway.",
            "title": "Transaction Id",
            "type": "integer"
          }
        },
        "required": [
          "transaction_id",
          "cp_id",
          "connector_id",
          "id_tag",
          "meter_start_wh",
          "started_at",
          "open"
        ],
        "title": "Transaction",
        "type": "object"
      },
      "TransactionDetail": {
        "example": {
          "connector_id": 1,
          "cp_id": "CP_BERLIN_001",
          "id_tag": "RFID_VALID_001",
          "meter_start_wh": 1000000,
          "meter_stop_wh": 1012345,
          "open": false,
          "request_id": "9b3c5d18-1f7c-4b6a-8e0e-5b9a3c4f2e10",
          "started_at": "2026-05-07T12:00:00+00:00",
          "stop_reason": "Local",
          "stopped_at": "2026-05-07T12:30:00+00:00",
          "transaction_id": 12345
        },
        "properties": {
          "connector_id": {
            "title": "Connector Id",
            "type": "integer"
          },
          "cp_id": {
            "title": "Cp Id",
            "type": "string"
          },
          "id_tag": {
            "description": "RFID / token used to authorize the session.",
            "title": "Id Tag",
            "type": "string"
          },
          "meter_start_wh": {
            "title": "Meter Start Wh",
            "type": "integer"
          },
          "meter_stop_wh": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meter Stop Wh"
          },
          "open": {
            "description": "True until StopTransaction lands.",
            "title": "Open",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "started_at": {
            "description": "ISO 8601 timestamp of the StartTransaction.",
            "title": "Started At",
            "type": "string"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Reason"
          },
          "stopped_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopped At"
          },
          "transaction_id": {
            "description": "OCPP 1.6 transaction id assigned by the gateway.",
            "title": "Transaction Id",
            "type": "integer"
          }
        },
        "required": [
          "transaction_id",
          "cp_id",
          "connector_id",
          "id_tag",
          "meter_start_wh",
          "started_at",
          "open",
          "request_id"
        ],
        "title": "TransactionDetail",
        "type": "object"
      },
      "TransactionListResponse": {
        "description": "`GET /api/v1/charge-points/{cp_id}/transactions`.",
        "example": {
          "request_id": "9b3c5d18-1f7c-4b6a-8e0e-5b9a3c4f2e10",
          "transactions": [
            {
              "connector_id": 1,
              "cp_id": "CP_BERLIN_001",
              "id_tag": "RFID_VALID_001",
              "meter_start_wh": 1000000,
              "meter_stop_wh": 1012345,
              "open": false,
              "started_at": "2026-05-07T12:00:00+00:00",
              "stop_reason": "Local",
              "stopped_at": "2026-05-07T12:30:00+00:00",
              "transaction_id": 12345
            }
          ]
        },
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "transactions",
          "next_cursor",
          "request_id"
        ],
        "title": "TransactionListResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Inbound REST surface (E3-7 / E3-8). Backend services read charger and transaction state under `/api/v1/charge-points`, `/api/v1/transactions`, etc., and dispatch OCPP commands via `/api/v1/charge-points/{cp_id}/commands/*`. All endpoints require a bearer token (`EVEYS_OCPP_REST_INBOUND_TOKENS`); `/api/v1/health` is exempt. Errors carry the closed-enum envelope `{error, error_code, request_id}` documented in [`docs/integration/02-gateway-rest-api.md`](https://github.com/eveys-mobility/OCPP/blob/main/docs/integration/02-gateway-rest-api.md).",
    "title": "eveys/ocpp gateway",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/charge-points/{cp_id}/transactions": {
      "get": {
        "operationId": "list_transactions_route_api_v1_charge_points__cp_id__transactions_get",
        "parameters": [
          {
            "in": "path",
            "name": "cp_id",
            "required": true,
            "schema": {
              "title": "Cp Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 10000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "id_tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id Tag"
            }
          },
          {
            "in": "query",
            "name": "open",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Open"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse",
                  "additionalProperties": true,
                  "title": "Response List Transactions Route Api V1 Charge Points  Cp Id  Transactions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad cursor or unparseable from/to timestamp."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List transactions for a charge point (cursor-paginated)",
        "tags": [
          "transactions"
        ]
      }
    },
...
    "/api/v1/transactions/{transaction_id}": {
      "get": {
        "operationId": "get_transaction_route_api_v1_transactions__transaction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDetail",
                  "additionalProperties": true,
                  "title": "Response Get Transaction Route Api V1 Transactions  Transaction Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown transaction_id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Single-transaction detail",
        "tags": [
          "transactions"
        ]
      }
    }
  }
}